backend/leaves/requirements.txt
# pipblinker==1.7.0
Celery==5.3.5
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
flask-marshmallow==1.2.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==1.0.0
packaging==23.2
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
SQLAlchemy==2.0.27
typing_extensions==4.10.0
Werkzeug==3.0.1
